package/waffle: Add the waffle package
The waffle package will be used for running piglit. By itself the
waffle package may be interesting to developers writing one off
applications or tests that link against waffle. It is more likely at
present that waffle will be used by the piglit package.

Ideally we'd want to use a release version of waffle, but there isn't
yet a waffle release that supports render nodes - which is a very
interesting features to have.

Cc: Chad Versace <chad.versace@linux.intel.com>
Signed-off-by: Ben Widawsky <ben@bwidawsk.net>