package/opengl/libgbm: Add the new opengl virtual
Several packages, most notably Waffle, are able to run GL workloads
without the traditional glx, or EGL support. GBM is the only
requirement. This can be incredibly useful when there is no display, or
the relevant drivers to bring up a full X windows system.

Someone with more knowledge could probably cut up the mesa3d.mk a bit
better. This is simply here so I can get the deps close to correct in
the upcoming package. You'll notice that EGL always implies GBM (which
was not a behavioral change)

(from the logs you seem to be the relevant mesa person)
Cc: Bernd Kuhls <bernd.kuhls@t-online.de>
Signed-off-by: Ben Widawsky <ben@bwidawsk.net>